+++ docs/CHANGELOG.md
@@ -11,6 +11,9 @@
 ## Python API
 * [Added] Declared compatibility with `jsonschema==4.*`.
 
+## Catalog, Lambdas
+* [Changed] `pkgpush` lambda now uses dedicated lambda for hashing files to push larger packages faster from catalog.
+
 ## CLI
 
 ## Catalog, Lambdas